Jenkins file updated

+++ Jenkinsfile
@@ -16,15 +16,23 @@ pipeline {
 
         stage('Build Docker Image') {
             steps {
-                sh "docker build -t ${DOCKER_IMAGE} ."
+                script {
+                    if (isUnix()) {
+                        sh "docker build -t ${DOCKER_IMAGE} ."
+                    } else {
+                        bat 'echo Skipping Docker build on Windows agent'
+                    }
+                }
             }
         }
 
         stage('Smoke Test') {
             steps {
-                sh '''
-                    docker run -d --rm --name ${APP_NAME}-test -p 8501:8501 ${DOCKER_IMAGE}
-                    python - <<'PY'
+                script {
+                    if (isUnix()) {
+                        sh '''
+                            docker run -d --rm --name ${APP_NAME}-test -p 8501:8501 ${DOCKER_IMAGE}
+                            python - <<'PY'
 import time
 import urllib.request
 import sys
@@ -40,15 +48,25 @@ for _ in range(30):
 print('Health check failed')
 sys.exit(1)
 PY
-                    docker stop ${APP_NAME}-test
-                '''
+                            docker stop ${APP_NAME}-test
+                        '''
+                    } else {
+                        bat 'echo Skipping Smoke Test on Windows agent'
+                    }
+                }
             }
         }
     }
 
     post {
         always {
-            sh 'docker rm -f ${APP_NAME}-test || true'
+            script {
+                if (isUnix()) {
+                    sh 'docker rm -f ${APP_NAME}-test || true'
+                } else {
+                    bat "docker rm -f ${APP_NAME}-test || exit 0"
+                }
+            }
         }
         success {
             echo "Pipeline succeeded: ${DOCKER_IMAGE}"
